+++ generic3g/ESMF_Subset.F90
@@ -9,9 +9,19 @@ module mapl3g_ESMF_Subset
    ! types
 
    use:: esmf, only: &
+        ESMF_VM, &
         ESMF_Clock, &
+        ESMF_Time, &
+        ESMF_TimeInterval, &
         ESMF_Config, &
+        ESMF_Geom, &
+        ESMF_Grid, &
+        ESMF_Mesh, &
+        ESMF_LocStream, &
+        ESMF_Xgrid, &
         ESMF_Field, &
+        ESMF_FieldBundle, &
+        ESMF_State, &
         ESMF_HConfig, &
         ESMF_HConfigIter, &
         ESMF_GridComp, &
@@ -27,11 +37,16 @@ module mapl3g_ESMF_Subset
         ESMF_METHOD_RUN, &
         ESMF_STATEINTENT_EXPORT, &
         ESMF_STATEINTENT_IMPORT, &
-        ESMF_SUCCESS
+        ESMF_SUCCESS, &
+        ESMF_CALKIND_GREGORIAN
         
    ! procedures
    use :: esmf, only: &
+        ESMF_TimePrint, &
+        ESMF_TimeSet, &
+        ESMF_CalendarSetDefault, &
         ESMF_HConfigAsStringMapKey, &
+        ESMF_HConfigAsString, &
         ESMF_HConfigCreate, &
         ESMF_HConfigCreateAt, &
         ESMF_HConfigDestroy, &
@@ -39,7 +54,21 @@ module mapl3g_ESMF_Subset
         ESMF_HConfigIterBegin, &
         ESMF_HConfigIterEnd, &
         ESMF_HConfigIterLoop, &
-        ESMF_HConfigGetSize
+        ESMF_HConfigGetSize, &
+        ESMF_VMGet, &
+        ESMF_VMGetCurrent, &
+        ESMF_ClockCreate, &
+        ESMF_ClockGet, &
+        operator(+), &
+        operator(-), &
+        operator(/), &
+        operator(*), &
+        operator(==), &
+        operator(/=), &
+        operator(<), &
+        operator(<=), &
+        operator(>), &
+        operator(>=)
 
     use :: esmf, only: &
          ESMF_InfoGetFromHost, &