+++ generic3g/MAPL_Generic.F90
@@ -1,3 +1,4 @@
+
 #include "MAPL_Generic.h"
 
 !---------------------------------------------------------------------
@@ -48,7 +49,8 @@ module mapl3g_Generic
    use esmf, only: ESMF_StateIntent_Flag, ESMF_STATEINTENT_INTERNAL
    use esmf, only: ESMF_KIND_I4, ESMF_KIND_I8, ESMF_KIND_R4, ESMF_KIND_R8
    use esmf, only: ESMF_KIND_R8, ESMF_KIND_R4
-   use esmf, only: ESMF_Time, ESMF_TimeInterval, ESMF_TimeIntervalGet, ESMF_Clock, ESMF_ClockGet
+   use esmf, only: ESMF_Time, ESMF_TimeInterval, ESMF_TimeIntervalGet, ESMF_Clock
+   use esmf, only: ESMF_ClockGet
    use esmf, only: ESMF_State, ESMF_StateItem_Flag, ESMF_STATEITEM_FIELD
    use esmf, only: operator(==)
    use mapl3g_hconfig_get
@@ -200,8 +202,7 @@ module mapl3g_Generic
    end interface MAPL_GridCompConnectAll
 
    interface MAPL_ClockGet
-      procedure :: clock_get
-      procedure :: ESMF_ClockGet
+      procedure :: clock_get_dt
    end interface MAPL_ClockGet
 
 contains
@@ -284,7 +285,7 @@ contains
       call MAPL_GridCompGetOuterMeta(gridcomp, outer_meta_, _RC)
 
       if (present(hconfig)) hconfig = outer_meta_%get_hconfig()
-      if (present(logger)) logger => outer_meta_%get_lgr()
+      if (present(logger)) logger => outer_meta_%get_logger()
       if (present(geom)) geom = outer_meta_%get_geom(_RC)
       if (present(grid)) then
          geom_ = outer_meta_%get_geom(_RC)
@@ -991,7 +992,7 @@ contains
       _UNUSED_DUMMY(unusable)
    end subroutine gridcomp_reexport
 
-   subroutine clock_get(clock, dt, rc)
+   subroutine clock_get_dt(clock, dt, rc)
       type(ESMF_Clock), intent(in) :: clock
       real(ESMF_KIND_R4), intent(out) :: dt ! timestep in seconds
       integer, optional, intent(out) :: rc
@@ -1004,6 +1005,6 @@ contains
       dt = real(seconds, kind=ESMF_KIND_R4)
 
       _RETURN(_SUCCESS)
-   end subroutine clock_get
+   end subroutine clock_get_dt
 
 end module mapl3g_Generic