+++ generic3g/OuterMetaComponent/run_child_by_name.F90
@@ -19,6 +19,8 @@ contains
       type(OuterMetaComponent), pointer :: child_meta
       logical :: found
       integer :: phase_idx
+      class(Logger), pointer :: lgr
+      character(:), allocatable :: this_name
 
       child = this%get_child(child_name, _RC)
       child_gc = child%get_gridcomp()
@@ -30,7 +32,11 @@ contains
          _ASSERT(found, "run phase: <"//phase_name//"> not found.")
       end if
 
+      lgr => this%get_logger()
+      this_name = this%get_name() ! workaround for gfortran
+      call lgr%debug('%a run child <%a~> (phase=%a~)', this_name, child_name, phase_name, _RC)
       call child%run(phase_idx=phase_idx, _RC)
+      call lgr%debug('  ... %a completed run child <%a~> (phase=%a~)', this_name, child_name, phase_name, _RC)
 
       _RETURN(_SUCCESS)
       _UNUSED_DUMMY(unusable)