+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add enumerators for RESTART setting
 - Add `add2export` to ACG3
 - Changed `MAPL_GridCompAddFieldSpec` and ACG3 to use new RESTART enum
+- Add validation for VariableSpec
 
 ### Changed
 
@@ -99,6 +100,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+
+- Added RC for Timer call in MAPL_Generic.F90
 
 ### Removed